Initial commit
While I have been working on this project for the last 6 months or so, I only just got around to putting it on GitHub. It has successfully been run on CP750 and CP850. A CP950 should have an effectively identical API as the CP850, so it should work. I only had the API for the JSD60, so that is completely untested, so I'm going to assume it doesn't work until I get a chance to work on it myself.